refactor(bridge): extract terminal outcome delivery
Anchors: src/services/discord/turn_bridge/mod.rs -> src/services/discord/turn_bridge/terminal_outcome_delivery.rs; maintenance baselines/docs updated for raw/prod counts.

Control-flow mapping: return=0, break=0, continue=0 targeting the enclosing spawn_turn_bridge scope, represented by TerminalOutcomeDeliveryOutcome::Completed. Internal-loop continue=1 remains internal. Error propagation: ?/.await? count=0.

Threaded locals: Context=24 fields, State=24 fields, Output=30 fields including the outcome. Delivery lease commit sites remain together with preserve_inflight_for_cleanup_retry and bridge_skip_holder_owns_inflight.

Gates: python3 scripts/generate_inventory_docs.py; python3 scripts/check_agent_maintenance_docs.py; bash scripts/ci-script-checks.sh; cargo fmt.
