+++ scripts/giant_file_registry.toml
@@ -487,6 +487,17 @@ deadline = "2026-08-31"
 decompose_issue = "#3038"
 
 [[entry]]
+# Created by #4230 S5 as the approved high-coupling terminal outcome delivery
+# extraction from turn_bridge/mod.rs. It remains over the giant threshold because
+# cancel, prompt-too-long, recovery retry, terminal delivery leases, controller
+# cutover, voice/TUI completion, dispatch completion, watcher-delivered marking,
+# and terminal readiness must stay together until the next decomposition stage.
+file = "src/services/discord/turn_bridge/terminal_outcome_delivery.rs"
+owner = "discord-relay"
+deadline = "2026-08-31"
+decompose_issue = "#4230"
+
+[[entry]]
 # Crossed the 1000-prod-LoC giant threshold in #3540: the durable pending
 # synthetic-turn-start worker (#3154) accreted the #3174/#3282/#3296/#3303/#3350
 # reconcile-marker chokepoints plus the #3540 no-evict post-abort queue promote.
